refactor(shell): extract leaf types module to resolve cyclic import
